refactor(chat): extract the inbox body to @endo/space-chat
Split the default 1:1 chat (inbox) into the standalone @endo/space-chat
package and a thin host wrapper, the first concrete cut of the
chat/channel family split.

- @endo/space-chat exports the pure `InboxRoot` Preact component (the
  message-list shell, all five message types, the reducer, the
  subscription, recipient filtering) authored with plain `preact` and
  importing its rendering utils from @endo/chat-kit. Confinement is the
  host's job.
- @endo/chat/inbox-component.js becomes the trusted wrapper: it resolves
  the mount, applies `renderConfined`, owns the scroll geometry (the
  component stays authority-free over the DOM), and returns `dispose`.
  Its entry signature is unchanged, so chat.js and every inbox test are
  untouched.

The package carries the same three pre-existing vnode-typing tsc
warnings the inbox always had (relocated, not new). Build and all five
inbox test suites pass.
