+++ packages/space-chat/index.js
@@ -1,0 +1,3 @@
+// @ts-check
+
+export { InboxRoot } from './src/inbox.js';
